+++ todolist/src/main/java/study/todolist/entity/todo/Status.java
@@ -1,4 +1,4 @@
-package study.todolist.entity;
+package study.todolist.entity.todo;
 
 public enum Status {
